+++ example/lib/page/sample_page.dart
@@ -3,6 +3,7 @@ import 'package:example/page/auto_load_page.dart';
 import 'package:example/page/basic_page.dart';
 import 'package:example/page/float_page.dart';
 import 'package:example/page/list_embed_page.dart';
+import 'package:example/page/manual_control_page.dart';
 import 'package:example/page/sliver_page.dart';
 import 'package:example/page/swiper_page.dart';
 import 'package:example/page/user_profile_page.dart';
@@ -36,6 +37,26 @@ class _SamplePageState extends State<SamplePage> {
                 }));
               },
               icon: Icon(Icons.format_list_bulleted,
+                color: Colors.orange,
+              ),
+            ),
+            Container(
+              width: double.infinity,
+              height: 0.5,
+              padding: EdgeInsets.only(left: 5.0, right: 5.0),
+              child: Container(
+                color: Colors.black12,
+              ),
+            ),
+            ListItem(
+              title: Translations.of(context).text("manualControl"),
+              describe: Translations.of(context).text("manualControlDescribe"),
+              onPressed: () {
+                Navigator.push(context,MaterialPageRoute(builder: (BuildContext context){
+                  return ManualControlPage();
+                }));
+              },
+              icon: Icon(Icons.gamepad,
                 color: Colors.orange,
               ),
             ),