+++ lib/ball_pulse_footer.dart
@@ -31,6 +31,8 @@ class BallPulseFooterState extends RefreshFooterState<BallPulseFooter> with Tick
   // 结束动画
   AnimationController _endController;
   Animation<double> _endAnimation;
+  // 是否开启动画
+  bool _isAnimation = false;
 
   // 三个球的大小
   double ballSize1 = 20.0;
@@ -44,6 +46,7 @@ class BallPulseFooterState extends RefreshFooterState<BallPulseFooter> with Tick
     _startController = new AnimationController(duration: const Duration(milliseconds: 400), vsync: this);
     _startAnimation = new Tween(begin: 6.0, end: 20.0).animate(_startController)
       ..addListener(() {
+        if (!mounted) return;
         setState(() {
           // 计算大小
           ballSize1 = 26.0 - _startAnimation.value;
@@ -65,6 +68,7 @@ class BallPulseFooterState extends RefreshFooterState<BallPulseFooter> with Tick
     _cycleController = new AnimationController(duration: const Duration(milliseconds: 800), vsync: this);
     _cycleAnimation = new Tween(begin: 6.0, end: 34.0).animate(_cycleController)
       ..addListener(() {
+        if (!mounted) return;
         setState(() {
           // 计算大小
           ballSize1 = _cycleAnimation.value <= 20.0 ? _cycleAnimation.value : 40.0 - _cycleAnimation.value;
@@ -81,10 +85,10 @@ class BallPulseFooterState extends RefreshFooterState<BallPulseFooter> with Tick
     _cycleAnimation.addStatusListener((status){
       if (status == AnimationStatus.completed) {
         _cycleController.reset();
-        if (this.refreshFooterStatus == RefreshFooterStatus.NO_LOAD) {
-          _endController.forward();
-        }else {
+        if (_isAnimation) {
           _cycleController.forward();
+        }else {
+          _endController.forward();
         }
       }
     });
@@ -92,6 +96,7 @@ class BallPulseFooterState extends RefreshFooterState<BallPulseFooter> with Tick
     _endController = new AnimationController(duration: const Duration(milliseconds: 400), vsync: this);
     _endAnimation = new Tween(begin: 6.0, end: 20.0).animate(_endController)
       ..addListener(() {
+        if (!mounted) return;
         setState(() {
           // 计算大小
           ballSize1 = _endAnimation.value;
@@ -114,16 +119,31 @@ class BallPulseFooterState extends RefreshFooterState<BallPulseFooter> with Tick
   @override
   Future onLoadStart() async {
     super.onLoadStart();
+    _isAnimation = true;
     _startController.forward();
   }
-
   // 正在加载
   @override
   Future onLoading() async {
     super.onLoading();
-    if (!_startController.isAnimating) {
+    if (!_isAnimation) {
+      _isAnimation = true;
       _startController.forward();
     }
+  }
+  // 加载结束
+  @override
+  Future onLoadEnd() async {
+    super.onLoadEnd();
+    _isAnimation = false;
+  }
+
+  @override
+  void dispose() {
+    _cycleController.dispose();
+    _startController.dispose();
+    _endController.dispose();
+    super.dispose();
   }
 
   @override